requirements_ocr.txt
# GOT-OCR 2.0 dependencies
transformers
torch
Pillow
pypdfium2
pypdf
